added todo about RHEL installs
problem will installing ICP on a fresh rhel box is the entitlements. We
should just wget the rpm and then let the rpm report that it cannot
install docker, otherwise there is no way to ensure we can install git
and docker.

The user will have to run the subscription-manager